Reorganized the back-end classes
