+++ src/main/java/com/depanneur_ste_helene/inventory_system/businesslayer/product/ProductService.java
@@ -1,8 +1,6 @@
-package com.depanneur_ste_helene.inventory_system.businesslayer;
+package com.depanneur_ste_helene.inventory_system.businesslayer.product;
 
-import com.depanneur_ste_helene.inventory_system.datalayer.Product;
-import com.depanneur_ste_helene.inventory_system.datalayer.ProductDTO;
-import org.springframework.stereotype.Service;
+import com.depanneur_ste_helene.inventory_system.datalayer.product.ProductDTO;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
