+++ src/main/java/com/depanneur_ste_helene/inventory_system/businesslayer/supplier/SupplierMapper.java
@@ -1,7 +1,7 @@
-package com.depanneur_ste_helene.inventory_system.businesslayer;
+package com.depanneur_ste_helene.inventory_system.businesslayer.supplier;
 
-import com.depanneur_ste_helene.inventory_system.datalayer.Supplier;
-import com.depanneur_ste_helene.inventory_system.datalayer.SupplierDTO;
+import com.depanneur_ste_helene.inventory_system.datalayer.supplier.Supplier;
+import com.depanneur_ste_helene.inventory_system.datalayer.supplier.SupplierDTO;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
