+++ src/main/java/com/depanneur_ste_helene/inventory_system/businesslayer/supplier/SupplierService.java
@@ -1,6 +1,6 @@
-package com.depanneur_ste_helene.inventory_system.businesslayer;
+package com.depanneur_ste_helene.inventory_system.businesslayer.supplier;
 
-import com.depanneur_ste_helene.inventory_system.datalayer.SupplierDTO;
+import com.depanneur_ste_helene.inventory_system.datalayer.supplier.SupplierDTO;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
