+++ src/main/java/com/depanneur_ste_helene/inventory_system/presentationlayer/ProductRESTController.java
@@ -1,8 +1,7 @@
 package com.depanneur_ste_helene.inventory_system.presentationlayer;
 
-import com.depanneur_ste_helene.inventory_system.businesslayer.ProductService;
-import com.depanneur_ste_helene.inventory_system.datalayer.Product;
-import com.depanneur_ste_helene.inventory_system.datalayer.ProductDTO;
+import com.depanneur_ste_helene.inventory_system.businesslayer.product.ProductService;
+import com.depanneur_ste_helene.inventory_system.datalayer.product.ProductDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
