+++ src/test/java/com/depanneur_ste_helene/inventory_system/CategoryServiceImplTests.java
@@ -1,10 +1,10 @@
 package com.depanneur_ste_helene.inventory_system;
 
-import com.depanneur_ste_helene.inventory_system.businesslayer.CategoryService;
-import com.depanneur_ste_helene.inventory_system.datalayer.Category;
-import com.depanneur_ste_helene.inventory_system.datalayer.CategoryCreateDTO;
-import com.depanneur_ste_helene.inventory_system.datalayer.CategoryDTO;
-import com.depanneur_ste_helene.inventory_system.datalayer.CategoryRepository;
+import com.depanneur_ste_helene.inventory_system.businesslayer.category.CategoryService;
+import com.depanneur_ste_helene.inventory_system.datalayer.category.Category;
+import com.depanneur_ste_helene.inventory_system.datalayer.category.CategoryCreateDTO;
+import com.depanneur_ste_helene.inventory_system.datalayer.category.CategoryDTO;
+import com.depanneur_ste_helene.inventory_system.datalayer.category.CategoryRepository;
 import com.depanneur_ste_helene.inventory_system.exceptions.AlreadyExistsException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
